Testcases added to support better test coverage
More testcases have been added for registering and
enrolling a user, and also for util functions to provide
better test coverage. Make file now checks to see
if there is 75% test coverage for each package. If not,
it will return appropriate message.

Code was also reorganized and directory structure was changed.

https://jira.hyperledger.org/browse/FAB-862

Change-Id: I82c0265fd3a2bbae4c2b9f2145234db818cbb8cb
Signed-off-by: Saad Karim <skarim@us.ibm.com>