Add support for cop client reenroll
This allows a user to renew a certificate by adding
support for the following command:
 cop client reenroll <copServerURL>
This is similar to the enroll command which is
 cop client enroll <enrollmentID> <enrollmentSecret> <copServerURL>
except that the reenroll uses token-based authentication rather than
basic auth.

More specifically, the authorization header of the request contains
a token which has a signature created from the private key of the
current enrollment certificate.  This allows a certificate to be
refreshed by the owner of the current certificate.  The body of a
reenroll request has a CSR just as the body of a enroll request.

Since there was a good bit in common between enroll and reenroll,
I rearranged some code to enable code sharing between the 2 paths.
In particular, some of the code from under the cli/client package
was moved to the lib package to allow it to be shared by enroll,
reenroll, or any other commands in the future.

See https://jira.hyperledger.org/browse/FAB-1147

Change-Id: I350a8bf8a700550e58b44beadc9a9af173cc2fee
Signed-off-by: Keith Smith <bksmith@us.ibm.com>