Group Prekey, Serial Number, and Max Enrollments
A prekey is generated for a group when a new group is created
and is stored in the groups table. Serial number and AKI has
been added to the users table and it will hold the serial number
and AKI for the current certificate for an enrolled. A configuration
option for repeated enrollment request by a user has been added. User
will be allowed a certain number of maximum enrollments which will
be specified in the server configuration file.

https://jira.hyperledger.org/browse/FAB-943
https://jira.hyperledger.org/browse/FAB-983

Change-Id: I34dff40a9eb8309da83395b242b24bb2e2c95873
Signed-off-by: Saad Karim <skarim@us.ibm.com>