Enforce validity period in COP for ECerts/TCerts
This change set verifies that a certificate has neither
expired nor been revoked.  The main change is a relatively
small change in auth.go which calls cfssl code to verify the
cert.

However, in the process of writing a test for this,
I had trouble setting an expiry time low enough without
affecting other test cases.  So I added another signing
profile to the COP server's config; search for "expiry"
in testdata/testconfig.json.

I realized that the enroll and reenroll commands do not
support using a non-default signing profile.
Therefore, I changed what goes across the network of an enroll
and reenroll request.  Prior to this change, it only sends
a CSR.  After this change, it sends JSON with a CSR and an
optional signing profile name (along with other optional fields).

I have notified the SDK folks of this change.

https://jira.hyperledger.org/browse/FAB-145

Change-Id: Idceeedadf9e5a42d995feee5bdba1353b685e9f1
Signed-off-by: Keith Smith <bksmith@us.ibm.com>