Enforce validity period in COP for ECerts/TCerts
This change set verifies that a certificate has neither
expired nor been revoked.  The main change for this
in auth.go which calls cfssl code to Verify the cert.

However, in the process of writing a test for this,
I had trouble because I couldn't set the normal expiry
too low since CFSSL code does some rounding down, etc.
So I as going to add another signing profile, something
that CFSSL supports, but realized in the process that we
didn't support this for enroll and reenroll.
The reason is that we were only sending a CSR across the wire.
I changed it to be like CFSSL and to send a
cfssl/signer.SignRequest across the wire, which contains a
CSR but also allows you to specify a non-default signing
profile and some other stuff to support HSM, etc.

So I had to add some fields to the idp.EnrollRequest and
idp.ReenrollRequest in idp.go; in particular, see
the final test case in cli/server/server_test.go which
sets a non-default signing profile.  The name of the profile
is "expiry" which you can see at the bottom of the
testdata/testconfig.json file, which is set to 1 second.
This not only allowed me to write a test case for this,
but also adds support for multiple signing profiles to
COP, something that will be useful in the future.

https://jira.hyperledger.org/browse/FAB-145

Change-Id: Idceeedadf9e5a42d995feee5bdba1353b685e9f1
Signed-off-by: Keith Smith <bksmith@us.ibm.com>