Patch cryptoutil to use bccsp to sign certs
Use bccsp to load a key and use that key to create a signer for cert
signing

The main code is fairly straightforward, except for being exceedently
verbose (all that error checking) and the need for extra indirection via
SKI file (BCCSP abstracts away the concept of keystore)

Probably the more contentios part is the test updates. Test material had
generated private keys and certs. Proper way is to generate them through
the bccsp interface so it could work with any bccsp. As a way to make do,
import the pem key into bccsp and update the SKI file, so that further
tests can work.

Change-Id: I13a3dcbb2aee53735e2c78d0930d5fc9493096c0
Signed-off-by: Volodymyr Paprotski <vpaprots@ca.ibm.com>