Adding TCert Library API
This is step 2 of adding TCert support to
the COP server.  In particular, the main
API is the GetBatch function in tcert.go.
All other functions support this main function.

In the next change set, we will integrate this
into the COP server by calling the GetBatch
function.

See https://jira.hyperledger.org/browse/FAB-876

Change-Id: I9eab8af2b6840865eaf4e4aa84458c2a1c535a1e
Signed-off-by: ASHUTOSH KUMAR <ashutosh_kumar@hotmail.com>