WIP: Improve docker build/experience
This change aligns docker build methods to be more generic,
more closely aligned with fabric.git, and allows the resulting
docker image to be more turnkey (it includes demo config/keys).

Known issues: cop is binding to localhost and thus not
properly exposing its services outside of the docker container.

Change-Id: Ifc7213bf7d3c74355e7013c09d5621b6cc06f54e
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>